Função de deletar produto criada, e o bug de crash do app ao deletar uma imagem no iphone e o bug da tela home duplicada

+++ lib/models/section.dart
@@ -101,7 +101,9 @@ class Section extends ChangeNotifier {
 
     for (final original in originalItems) {
       try {
-        if (!items.contains(original)) {
+        if (!items.contains(original)
+            && (original.image as String).contains('firebase'))
+        {
           final ref =
           await storage.getReferenceFromUrl(original.image as String);
           await ref.delete();
@@ -120,13 +122,15 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = await storage.getReferenceFromUrl(
-            item.image as String
-        );
-        await ref.delete();
-      // ignore: empty_catches
-      } catch (e){}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(
+              item.image as String
+          );
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e) {}
+      }
     }
   }
